refactor(headless-crawler): 階層の更新 - ヘッドレスブラウザの操作なので、browserディレクトリ内に、階層的にディレクトリを配置するのが自然かなと
