+++ packages/headless-crawler/lib/domains/core/browser/interactions/extraction/jobDetail/error.ts
