+++ packages/headless-crawler/lib/domains/core/browser/interactions/index.ts
@@ -1,0 +1,4 @@
+export * from "./element-action";
+export * from "./extraction";
+export * from "./navigation";
+export * from "./form-filling";
