+++ packages/headless-crawler/lib/domains/core/browser/interactions/navigation/error.ts
